Enforce Subject Management user permissions
- also add empty table message

+++ app/Http/Controllers/SubjectController.php
@@ -7,11 +7,14 @@ use App\Models\Proficiency;
 use App\Models\Subject;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Gate;
 
 class SubjectController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-subjects');
+
         $subjects = Subject::with(['accommodations:id,name', 'proficiencies:id,name'])->get();
         $accommodations = Accommodation::orderBy('name')->get(['id', 'name']);
         $proficiencies = Proficiency::orderBy('name')->get(['id', 'name']);
@@ -21,6 +24,8 @@ class SubjectController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create-subjects');
+
         $validated = $request->validate([
             'name' => 'required|unique:subjects,name|max:255',
             'code' => 'nullable|string|unique:subjects,code|max:255',
@@ -42,6 +47,8 @@ class SubjectController extends Controller
 
     public function update(Request $request, Subject $subject)
     {
+        Gate::authorize('edit-subjects');
+
         $validated = $request->validate([
             'name' => 'required|max:255|unique:subjects,name,' . $subject->id,
             'code' => 'nullable|string|max:255|unique:subjects,code,' . $subject->id,
@@ -64,6 +71,8 @@ class SubjectController extends Controller
 
     public function destroy(Subject $subject)
     {
+        Gate::authorize('delete-subjects');
+        
         try {
             $subject->delete();
             return back()->with('success', 'Subject Deleted Successfully!');
